+++ package.json
@@ -32,8 +32,8 @@
     "react-router": "^6.0.0-beta.1",
     "react-router-dom": "^6.0.0-beta.1",
     "react-scripts": "4.0.3",
-    "react-visibility-sensor": "^5.1.1",
     "spotify-web-api-js": "^1.5.1",
+    "uuid": "^8.3.2",
     "web-vitals": "^1.0.1",
     "workbox-background-sync": "^5.1.3",
     "workbox-broadcast-update": "^5.1.3",
