--- src/components/Track.css
@@ -1,3 +1,0 @@
-.highlight {
-  border: 2px solid white;
-}
