+++ src/components/Track.js
@@ -1,10 +1,5 @@
-import { PureComponent, Suspense, useState } from 'react';
-import { View, Spinner } from './Platform';
-import VisibilitySensor from 'react-visibility-sensor';
-
-try {
-  require('./Track.css');
-} catch (e) { }
+import React, { PureComponent, Suspense, useState } from 'react';
+import { View, Spinner, IFrame } from './Platform';
 
 function createResource(pending) {
   let error, response;
@@ -21,39 +16,28 @@ function createResource(pending) {
 // Avoid unncessary re-renders
 class SpotifyTrackWidget extends PureComponent {
   render() {
-    const { track_id, highlight, resource, testing } = this.props;
+    const { track_id, resource } = this.props;
     const iframe = (resource) ? resource.read() : null;
 
     return (
-      <VisibilitySensor offset={{ top: -5 * 80, bottom: -5 * 80 }}>
-        {({ isVisible }) =>
-          <iframe
-            className={highlight ? 'highlight' : ''}
-            title={track_id}
-            // src={`https://open.spotify.com/embed/track/${track_id}`}
-            srcdoc={(isVisible || testing) ? Buffer.from(iframe.data, 'base64') : ''}
-            width='100%'
-            height='80'
-            frameBorder='0'
-            allowtransparency='true'
-            allow='encrypted-media'
-          />
-        }
-      </VisibilitySensor>
+      <IFrame
+        title={track_id}
+        width='100%'
+        height={80}
+        srcdoc={Buffer.from(iframe.data, 'base64').toString()}
+      />
     );
   }
 }
 
 export default function Track({
-  track_id,
-  highlight = false,
-  testing = false
+  track_id
 }) {
   if (typeof Track.cache == 'undefined') {
     Track.cache = {};
   }
 
-  const [ resource, _ ] = useState((track_id in Track.cache) ? Track.cache[track_id] :
+  const [resource, ] = useState((track_id in Track.cache) ? Track.cache[track_id] :
     createResource(new Promise(resolves => {
       fetch(`${process.env.REACT_APP_API_URL}/widget?track_id=${track_id}`)
         .then(response => (response.status === 200) ? response.text() : '')
@@ -78,9 +62,7 @@ export default function Track({
     } >
       <SpotifyTrackWidget
         track_id={track_id}
-        highlight={highlight}
         resource={resource}
-        testing={testing}
       />
     </Suspense>
   );
